[cleanup] Moved readers to approprate demuxers folder
