+++ src/CMakeLists.txt
@@ -4,10 +4,7 @@ set(SOURCES
   oscompat.cpp
   Session.cpp
   Stream.cpp
-  TSReader.cpp
   aes_decrypter.cpp
-  ADTSReader.cpp
-  WebmReader.cpp
 )
 
 set(HEADERS
@@ -16,11 +13,8 @@ set(HEADERS
   oscompat.h
   Session.h
   Stream.h
-  TSReader.h
   aes_decrypter.h
-  ADTSReader.h
   Iaes_decrypter.h
-  WebmReader.h
 )
 
 add_dir_sources(SOURCES HEADERS)
@@ -28,6 +22,7 @@ add_dir_sources(SOURCES HEADERS)
 add_subdirectory(codechandler)
 add_subdirectory(common)
 add_subdirectory(decrypters)
+add_subdirectory(demuxers)
 add_subdirectory(parser)
 add_subdirectory(samplereader)
 add_subdirectory(utils)
