+++ src/demuxers/CMakeLists.txt
@@ -1,0 +1,13 @@
+set(SOURCES
+  ADTSReader.cpp
+  TSReader.cpp
+  WebmReader.cpp
+)
+
+set(HEADERS
+  ADTSReader.h
+  TSReader.h
+  WebmReader.h
+)
+
+add_dir_sources(SOURCES HEADERS)
