+++ src/main.cpp
@@ -8,10 +8,7 @@
 
 #include "main.h"
 
-#include "ADTSReader.h"
 #include "Stream.h"
-#include "TSReader.h"
-#include "WebmReader.h"
 #include "samplereader/ADTSSampleReader.h"
 #include "samplereader/FragmentedSampleReader.h"
 #include "samplereader/SubtitleSampleReader.h"
